session11_thumbnail/requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-appconf==1.0.3
django-crispy-forms==1.7.2
django-imagekit==4.0.2
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.1
pilkit==2.0
Pillow==6.0.0
psycopg2-binary==2.8.2
python3-openid==3.1.0
pytz==2018.9
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.25.2
whitenoise==4.1.2
